Refactoring the logic of finding the active rule from CoveritySensor into CoverityRuleUtil.
